requirements.txt
transformers[torch]
pillow==10.3.0
streamlit==1.34.0
requests==2.31.0
tokenizers==0.13.3